+++ test/test-compare-pcqm.sh
@@ -50,4 +50,14 @@ then
 		 > ${TMP}/${OUT}.txt 2>&1
 	grep -iF "error" ${TMP}/${OUT}.txt
 
+	OUT=compare_pcqm_longdress_near_lossless
+	$CMD  \
+		compare --mode pcqm  \
+		--inputModelA ${EXTDATA}/longdress_fr1051.obj \
+		--inputMapA  ${EXTDATA}/longdress_fr1051.png \
+		--inputModelB ${EXTDATA}/longdress_fr1051_tfan.obj \
+		--inputMapB  ${EXTDATA}/longdress_fr1051.png \
+		--topologyFile ${EXTDATA}/longdress_fr1051_tfan_topo.txt \
+		 > ${TMP}/${OUT}.txt 2>&1
+	grep -iF "error" ${TMP}/${OUT}.txt
 fi